requirements.txt
Keras == 2.4.3
Pillow == 7.0.0
albumentations == 0.1.12
efficientnet == 1.1.1
h5py == 2.10.0
imageio == 2.4.1
imgaug == 0.2.9
numpy == 1.19.5
opencv_python == 4.1.2.30
pandas == 1.1.5
scikit_image == 0.16.2
scipy == 1.4.1
seaborn == 0.11.1
sklearn == 0.0
tqdm == 4.41.1
segmentation_models == 1.0.1
tensorflow == 2.1.0
